### 2023-07-10 #5641 @editorialbot
[PRE REVIEW]: inAccessMod: An R package to automate data downloading and processing for AccessMod
**Submitting author:** @ptimoner (<a href="http://orcid.org/0000-0003-4757-4928">Pablo Timoner</a>)
**Repository:** https://github.com/unige-geohealth/inAccessMod
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
