### 2023-08-21 9b12cc3
new field for filtering process
