+++ R/hf_best_cov.R
@@ -185,6 +185,7 @@ hf_best_cov <- function (workDir,
   
   # Keep only the catchments that are different for the main algorithm
   tempCatchUnique <- tempCatch[!duplicated(tempCatch$geometry), ]
+  tempCatchUnique$totalpop0 <- tempCatchUnique$totalpop
   
   # Main algorithm
   message("Main algorithm...")
@@ -224,12 +225,20 @@ hf_best_cov <- function (workDir,
       if (length(notComplete) == 0) {
         # If all complete, no constraint
         indMax <- which(tempCatchUnique$totalpop == max(tempCatchUnique$totalpop))
+        if (length(indMax) > 1) {
+          # [1] In case they had the exact same initial covered population (but not identical catchment)
+          indMax <- indMax[which(tempCatchUnique$totalpop0[indMax] == max(tempCatchUnique$totalpop0[indMax]))][1]
+        }
       } else {
         # Which are the admin that have priority
         tempAdmin <- sf::st_drop_geometry(tempCatchUnique[, adminColName])[, 1]
         validRows <- tempAdmin %in% notComplete
         # Look for the max among them
         indMax <- which(tempCatchUnique$totalpop == max(tempCatchUnique$totalpop[validRows]))
+        if (length(indMax) > 1) {
+          # [1] In case they had the exact same initial covered population (but not identical catchment)
+          indMax <- indMax[which(tempCatchUnique$totalpop0[indMax] == max(tempCatchUnique$totalpop0[indMax]))][1]
+        }
         # Increment the count for the selected admin
         selAdmin <- sf::st_drop_geometry(tempCatchUnique[, adminColName])[indMax, ]
         hfCounts$count[hfCounts$admin == selAdmin] <- hfCounts$count[hfCounts$admin == selAdmin] + 1
@@ -237,17 +246,16 @@ hf_best_cov <- function (workDir,
     } else {
       # Max with no constraint
       indMax <- which(tempCatchUnique$totalpop == max(tempCatchUnique$totalpop))
-    }
-    # Usually only one catchment (same population in two non-identical catchments is very unlikely)
-    # But we never knows.
-    for (j in 1:length(indMax)) {
-      # Increment of i
-      i <- i + 1
-      finalTable[i, "Facility name"] <- sf::st_drop_geometry(tempCatchUnique[indMax, catchHfColName])[j, 1]
-      finalTable[i, "Population covered"] <- tempCatchUnique$totalpop[indMax][j]
-      if (adminCheck) {
-        finalTable[i, "Region"] <- sf::st_drop_geometry(tempCatchUnique[indMax, adminColName])[j, 1]
+      if (length(indMax) > 1) {
+        # [1] In case they had the exact same initial covered population (but not identical catchment)
+        indMax <- indMax[which(tempCatchUnique$totalpop0[indMax] == max(tempCatchUnique$totalpop0[indMax]))][1]
       }
+    }
+    i <- i + 1
+    finalTable[i, "Facility name"] <- sf::st_drop_geometry(tempCatchUnique[indMax, catchHfColName])[1, 1]
+    finalTable[i, "Population covered"] <- tempCatchUnique$totalpop[indMax]
+    if (adminCheck) {
+      finalTable[i, "Region"] <- sf::st_drop_geometry(tempCatchUnique[indMax, adminColName])[1, 1]
     }
     # Get catchment(s) with max population
     top <- tempCatchUnique[indMax, ]
@@ -276,42 +284,36 @@ hf_best_cov <- function (workDir,
     }
   }
   close(bar)
-  
-  # Calculation of outputs
   colNamesFT <- colnames(finalTable)
[…]
