### 2023-09-23 #5641 @arfon
@gkthiruvathukal – just a heads up that @melissawm is stepping down from the JOSS editorial team. I'm happy to pick this one up!
