### 2023-09-23 #5641 @arfon
@mponce0 @joseph-palmer @ConnorDonegan – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is inAccessMod: An R package to automate data downloading and processing for AccessMod: https://github.com/unige-geohealth/inAccessMod

As previous authors of somewhat similar packages, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
