### 2023-09-24 #5641 @ConnorDonegan
Hi @arfon, I'll be glad to review.

@ptimoner, it looks like the file initiate2.png was renamed to initiate2.PNG, causing this issue on installation for me:

```
> devtools::install_github("unige-geohealth/inAccessMod", build_vignettes = TRUE)
Downloading GitHub repo unige-geohealth/inAccessMod@HEAD
── R CMD build ─────────────────────────────────────────────────────────────────
   checking for file ‘/tmp/RtmptTu4Pn/remotes1c2476026e73f/unige-geohealth-inAcc✔  checking for file ‘/tmp/RtmptTu4Pn/remotes1c2476026e73f/unige-geohealth-inAccessMod-d2aef72/DESCRIPTION’
─  preparing ‘inAccessMod’:
E  creating vignettes (4.3s)
   --- re-building ‘Tutorial.Rmd’ using rmarkdown
   File initiate2.png not found in resource path
   Error: processing vignette 'Tutorial.Rmd' failed with diagnostics:
   pandoc document conversion failed with error 99
   --- failed re-building ‘Tutorial.Rmd’

   SUMMARY: processing the following file failed:
     ‘Tutorial.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error: Failed to install 'inAccessMod' from GitHub:
  ! System command 'R' failed
```
