### 2023-09-24 a9c63f4
core.ignorecase false
