### 2023-09-24 #5641 @ptimoner
Dear @ConnorDonegan,

First of all thanks a lot for accepting to review our package.

I renamed the files that had the extension 'PNG' instead of 'png'. Let me know if you still get an issue when installing the package.

Best regards,

Pablo
