### 2023-10-16 unige-geohealth/inAccessMod#3 @ConnorDonegan
initiate_project issue
Hi @ptimoner

I'm having trouble working through the vignette. It looks like there is a capitalization of country name that was not expected or changed at some point.

I run this:

```
mainPath <- getwd()
initiate_project(mainPath)
```

and select a country, and then
```
Select the country
...
...
## Do you want to add another input folder (type 1 or 2)?

## 1: YES
## 2: NO

## Selection: 2
## Error in writeLines(c(paste0("COUNTRY:", countryOriginalName), paste0("ISO:",  :
##   cannot open the connection
## In addition: Warning message:
## In writeLines(c(paste0("COUNTRY:", countryOriginalName), paste0("ISO:",  :
##   cannot open file '/home/connor/peer-review/JOSS/Switzerland/data/config.txt': No such file or directory
```
The path that was created for me is: `/home/connor/peeer-review/JOSS/SWITZERLAND/data` with caps on country name.

@https://github.com/openjournals/joss-reviews/issues/5879
