### 2023-10-17 unige-geohealth/inAccessMod#3 @ptimoner
Indeed, it appears that there was an unnecessary capitalization of the country name. Could you please attempt it again? If the issue is resolved, kindly close it. Thanks !

@https://github.com/openjournals/joss-reviews/issues/5879
