+++ vignettes/Tutorial.Rmd
@@ -19,7 +19,7 @@ library(inAccessMod)
 ```
 
 ### Description
-This package allows us to easily prepare all the required layers for AccessMod. A proper folder structure is created in order to manage multi-temporal data and/or multiple HeRAMS analysis scenarios. While the functions to process health facility tables are specifically designed to handle HeRAMS data, the other ones can be used for any other project. The downloading, cropping, masking, resampling and exporting processes are automated to a large degree, making the preparation of the inputs quick and straightforward.
+This package allows us to easily prepare all the required layers for AccessMod. A proper folder structure is created in order to manage multi-temporal data and/or multiple analysis scenarios for The Health Resources and Services Availability Monitoring System (HeRAMS). While the functions to process health facility tables are specifically designed to handle HeRAMS data, the other ones can be used for any other project. The downloading, cropping, masking, resampling and exporting processes are automated to a large degree, making the preparation of the inputs quick and straightforward.
 
 ### Installation
 First, we make sure that the devtools package installed on our computer.
@@ -59,10 +59,11 @@ Now, that the project is initiated we can use a set of functions that allow to d
 
 ```{r copy_input, eval=FALSE}
 country <- "Switzerland"
-copy_input(mainPath, country, "C:/example/boundaries")
+inputPath <- system.file("extdata", "swiss_boundaries_example", package="inAccessMod")
+copy_input(mainPath, country, inputPath)
 ```
 
-We are then asked to select the data we want to copy.
+We are then asked to select the data we want to copy. In this case, as the folder contains shapefile of the boundaries, we would select index: 4
 
 ![](copy_inputs.png)
 
@@ -76,10 +77,10 @@ download_boundaries(mainPath, country, adminLevel = 1, type = "gbOpen", alwaysDo
 
 ### Subset regions
 
-In case we would like to work only on specific country regions, we use the `subset_regions` function. It subsets the boundary shapefile based on region names included in the shapefile attribute table. We are first asked to select the column that contains the region names, and then the region that we want to keep. A new boundary shapefile will be created. The function only requires the project root path and the country folder name.
+In case we would like to work only on specific country regions, we use the `subset_regions` function. It subsets the boundary shapefile based on region names included in the shapefile attribute table. We are first asked to select the column that contains the region names, and then the region that we want to keep. A new boundary shapefile will be created. The function only requires the project root path and the country folder name. In the example below we also set the parameter `mostRecent` to TRUE so it will automatically load the most recent downloaded shapefile. We will be asked to select the column that we would like to use for subsetting the regions, let's select the column 1 that refers to the names of the regions. A list of the region names will be displayed so we can select the regions. Let's select the regions "Genève" and "Vaud", by indicating their indices separated by whitespaces: 7 8 11 13 23 24 
 
 ```{r subset_regions, eval=FALSE}
-subset_regions(mainPath, country)
+subset_regions(mainPath, country, mostRecent = TRUE)
 ```
 
 ### Set the projected coordinate reference system
@@ -100,9 +101,9 @@ The function also has projected the boundary shapefile and we can observe how th
 
 ### Filter the health facilities and create a point shapefile (HeRAMS table)
 
[…]
