+++ R/initiate_project.R
@@ -7,6 +7,11 @@
 #' @details The final structure arises when downloading and processing the data with the corresponding functions,
 #' and it allows multiple 'raw' inputs and multiple 'processed' outputs for each input. This can be useful when 
 #' performing different analyses for the same country (e.g. when we have updated data).
+#' @examples
+#' # Replace workDir with the actual path to your working directory
+#' \dontrun{
+#' mainPath <- "workDir"
+#' initiate_project(mainPath)}
 #' @export
 initiate_project <- function (mainPath) {
   if (!is.character(mainPath)) {
