+++ R/set_projection.R
@@ -15,6 +15,17 @@
 #' \code{suggest_top_crs} and the \code{suggest_crs}, respectively, from the \pkg{crsuggest} package.
 #' Both function work by analyzing the extent of the spatial dataset and comparing it to the area extents
 #' in the EPSG's coordinate reference system database.
+#' @examples
+#' # Replace workDir with the actual path to your working directory
+#' \dontrun{
+#' mainPath <- "workDir"
+#' intiate_project(mainPath, country)}
+#' 
+#' # Replace myCountry with the country name you are working on (workDir subfolder)
+#' \dontrun{
+#' country <- "myCountry"
+#' download_boundaries(mainPath, country, adminLevel = 1, type = "gbOpen", alwaysDownload = TRUE)
+#' set_projection(mainPath, country, mostRecent = TRUE, alwaysSet = TRUE, bestCRS = TRUE)}
 #' @export
 set_projection <- function (mainPath, country, mostRecent = FALSE, alwaysSet = FALSE, bestCRS = FALSE) {
   if (!is.character(mainPath)) {
