+++ man/set_HeRAMS_table_parameters.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/set_herams_table_parameters.R
+% Please edit documentation in R/set_HeRAMS_table_parameters.R
 \name{set_HeRAMS_table_parameters}
 \alias{set_HeRAMS_table_parameters}
 \title{Set HeRAMS table parameters}
