### 2023-10-23 0b72d46
data for hf_best_cov example
