+++ R/hf_best_cov.R
@@ -3,7 +3,8 @@
 #' Select the health facilities that offer the best population coverage using their 
 #' catchments and a raster of population. This function also offers the possibility to ensure that N 
 #' facilities are selected for each administrative unit.
-#' @param workDir character; working directory that contains the inputs
+#' @param inputFolder character; working directory that contains the inputs
+#' @param outputFolder character; the path to the output folder. If NULL, an output folder will be created within the inputFolder.
 #' @param catchShp character; name of the catchment shapefile WITHOUT extension
 #' @param popRaster character; name of the population raster file WITH extension
 #' @param catchHfColName character; name of the health facility name column in the catchment shapefile
@@ -21,7 +22,8 @@
 #' It repeats the process until the number of facilities to be selected is reached.
 #' @export
 
-hf_best_cov <- function (workDir, 
+hf_best_cov <- function (inputFolder,
+                        outputFolder,
                         catchShp, 
                         popRaster, 
                         catchHfColName, 
@@ -35,8 +37,8 @@ hf_best_cov <- function (workDir,
     
   #Loading the inputs
   message("Loading the inputs...")
-  if (!dir.exists(workDir)) {
-    stop(paste(workDir, "does not exist."))
+  if (!dir.exists(inputFolder)) {
+    stop(paste(inputFolder, "does not exist."))
   }
   if (!is.numeric(nTot)) {
     stop("nTot must be 'numeric'")
@@ -44,12 +46,12 @@ hf_best_cov <- function (workDir,
   if (!is.logical(adminCheck)) {
     stop("adminCheck must be 'logical'")
   }
-  tempCatch <- file.path(workDir, paste0(catchShp, ".shp"))
+  tempCatch <- file.path(inputFolder, paste0(catchShp, ".shp"))
   if (!file.exists(tempCatch)) {
     stop(paste(tempCatch, "does not exist."))
   }
   tempCatch <- sf::st_read(tempCatch, quiet = TRUE)
-  pop <- file.path(workDir, popRaster)
+  pop <- file.path(inputFolder, popRaster)
   if (!file.exists(pop)) {
     stop(paste(pop, "does not exist."))
   }
@@ -73,7 +75,7 @@ hf_best_cov <- function (workDir,
     if (is.null(hfShp)) {
       stop("If adminCheck = TRUE, hfShp is required.")
     }
-    hf <- file.path(workDir, paste0(hfShp,".shp"))
+    hf <- file.path(inputFolder, paste0(hfShp,".shp"))
     hf <- sf::st_read(hf, quiet = TRUE)
     if (!hfHfColName %in% colnames(hf)) {
       stop(paste(hfHfColName, "is not a valid column name in the facility shapefile."))
@@ -83,7 +85,7 @@ hf_best_cov <- function (workDir,
     if (!all(c(test1, test2))) {
       stop("Discrepancy between facility names in health facility and catchment shapefiles.")
     }
-    admin <- file.path(workDir, paste0(adminShp,".shp"))
+    admin <- file.path(inputFolder, paste0(adminShp,".shp"))
     admin <- sf::st_read(admin, quiet = TRUE)
     if (!adminColName %in% colnames(admin)) {
       stop(paste(hfHfColName, "is not a valid column name in the facility shapefile."))
@@ -122,29 +124,46 @@ hf_best_cov <- function (workDir,
   tempCatch$totalpop <- exactextractr::exact_extract(pop, tempCatch, "sum", progress = FALSE)
   # Catchment having same population and potentially identical (verified using st_equals)
[…]
