### 2023-10-23 unige-geohealth/inAccessMod#4 @ptimoner
Dear @ConnorDonegan ,

Thanks a lot for your time in reviewing my package and for all your suggestions. I followed all of them as I believed they could really improve the support to the user.

- I modified the vignette, and it is now possible to move from "initiate_project" (first step) to "compile_processed_data" (last step) just following the Tutorial, using the same example project across the different functions.
- I added example codes to the help pages.
- I added example data within the package for the additional functions (multi_ts and hf_best_cov), the ones that can be run outside the main pipeline, and that can be useful for AccessMod users. This way, the user can test the functions even they don't have their own inputs.
- I added the instructions to the README file so users know how to obtain the vignette.
- I spelled out HeRAMS in the Tutorial.
- I added the link to AccessMod page (where there is all the information for its installation and use), in both the README file and the Tutorial.

I hope these changes meet your expectations, and the package is now more user-friendly.

Thank you again for your valuable insights,

Best regards,

Pablo

https://github.com/openjournals/joss-reviews/issues/5879
