### 2023-10-24 unige-geohealth/inAccessMod#4 @ConnorDonegan
Thanks Pablo, I look forward to working through the new material (within the coming week or two).
