### 2023-11-08 #5879 @mponce0
@arfon I'm done with the review.

I'm including some additional elements for the authors that could help address some of the missing points:

# General checks
> * Contribution and authorship: Has the submitting author (@ptimoner) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

@ptimoner appears to be the main contributor to the repository.
However, it is not clear the role of each of the authors in the paper/package, a statement of contribution can help with this.

>  Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

 The package appears to build upon a well established open source tool in the field modelling health care access.
 It looks to be a sort of wrapper to provide quick access to the data and models from this former tool.

> Human and animal research: If the paper contains original data research on humans subjects or animals, does it comply with JOSS's human participants research policy and/or animal research policy? If the paper contains no such data, please check this item.

As the paper is based on health care access data, although the package itself does not contain it but gives access to it... I'm not sure whether we should look deeper into this?

# Functionality
>   Installation: Does installation proceed as outlined in the documentation?

Not specified in the paper, but is an R package so it should follow the usual standards, still borth menitoning it.

> Functionality: Have the functional claims of the software been confirmed?

Can not verify that, as no examples are provided.

# Documentation
> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Partially, nothing is mentioned in the manuscript but instructions can be found in the package github repository.

>  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

No examples are included, I would strongly suggest to do so.

>  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Partially, a very high level overview is described in the manuscript.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

No

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

No specifications neither about contributions nor support are provided.

# Software paper
> State of the field: Do the authors describe how this software compares to other commonly-used packages?

No comparison is done with other packages nor mentioned whether there is a similar package other than the "mother" tool used to harvest the data from.

---
# Further recommendations:

- submit to CRAN

 - add examples

---
