### 2023-11-09 #5879 @ptimoner
@mponce0  Thanks a lot for taking time to review our package. Regarding the documentation and functionality sections, I just wanted to mention that there is vignette available. Following @ConnorDonegan recommendations (https://github.com/unige-geohealth/inAccessMod/issues/4), we added the instructions to the README file so users know how to obtain the vignette, we modified the vignette, and it is now possible to move from "initiate_project" (first step) to "compile_processed_data" (last step) just following the Tutorial, using the same example project across the different functions. We also added example codes to the help pages, and we added example data within the package for the additional functions (multi_ts and hf_best_cov), the ones that can be run outside the main pipeline, and that can be useful for AccessMod users. This way, the user can test the functions even they don't have their own inputs. I hope that this clarifies the main concern raised in your review.
Thanks again,
Best,
Pablo
