### 2023-11-11 #5879 @mponce0
Hi Pablo @ptimoner , thanks for the clarification, this helps.
However, I wonder are you planning on submitting the packages to CRAN or bioconductor?
If the answer is not maybe it would be worth it presenting the tutorial and examples as a main addition to the repo, vignettes make a lot of sense for R package but really useful when hosted on the main R repositories as are web-rendered and provide easy access for the users... in your case you could do the same by just offering a markdown or PDF version in your repository again unless your plans are to submit to CRAN/bioconductor.

WRT installation instructions, I have been having some issues installing the package.
For instance,
```R
-----------------------------------
ERROR: this R is version 4.1.2, package 'inAccessMod' requires R >=  4.1.3
-----------------------------------
ERROR: package installation failed
Error: Failed to install 'inAccessMod' from GitHub:
! System command 'R' failed
In addition: There were 13 warnings (use warnings() to see them)
```
as indicated in the instructions section, this type of things and other dependencies should be mentioned.

Similarly, when trying to install the package in R v 4.3.2, some missing dependencies are making the installation fail, e.g.
```R
ERROR: dependencies ‘crsuggest’, ‘exactextractr’, ‘gdalUtils’, ‘osmextract’, ‘rgdal’, ‘rgeoboundaries’, ‘rgeos’, ‘sf’ are not available for package ‘inAccessMod’
```
