+++ TUTORIAL.md
@@ -1,0 +1,575 @@
+    library(inAccessMod)
+
+### Description
+
+This package allows us to easily prepare all the required layers for
+AccessMod. A proper folder structure is created in order to manage
+multi-temporal data and/or multiple analysis scenarios for The Health
+Resources and Services Availability Monitoring System (HeRAMS). While
+the functions to process health facility tables are specifically
+designed to handle HeRAMS data, the other ones can be used for any other
+project. The downloading, cropping, masking, resampling and exporting
+processes are automated to a large degree, making the preparation of the
+inputs quick and straightforward. More information on AccessMod
+installation and use can be found here: <https://www.accessmod.org/>
+
+### Installation
+
+First, we make sure that the devtools package installed on our computer.
+
+    if (!require("devtools")) install.packages("devtools")
+
+Then we use its `install_github` function to download and install the
+`inAccessMod` package from its github repository.
+
+    devtools::install_github("https://github.com/unige-geohealth/inAccessMod", build_vignettes = TRUE)
+
+Once it is installed, we can load the package in order to make its
+functions directly available.
+
+    library(inAccessMod)
+
+### Initiate Project
+
+With the `initiate_project` function, we can select the country and
+automatically, the ISO 3166-1 alpha-3 country code is stored in a
+config.txt file and the directory main structure for the project is
+created. This function also creates a log.txt file that will record and
+track the main operations related to the project. We will see that the
+final structure arises when downloading and processing the data with the
+corresponding functions, and it allows multiple ’raw’ inputs and
+multiple ’processed’ outputs for each input. This can be useful when
+performing different analyses for the same country (e.g. when we have
+updated data) or using different processing parameters. To initiate the
+project, the only parameter we need to provide is the path where the
+project, namely the country folder will be created. Let’s run the
+function and select “Switzerland” for this tutorial.
+
+    # Replace workDir with the actual path to your working directory
+    mainPath <- getwd()
+    initiate_project(mainPath)
+
+![](initiate.png)
+
+As indicated in the console, some input folders are created by default.
+We are asked if we want to add another input folder. Let’s say that it
+is not necessary. This is how the new directories look like in a
+tree-like format. Further, we will see how this tree can grow with the
+downloading and the processing of inputs.
+
+![](initiate2.png)
+
+### Import inputs
+
+Now, that the project is initiated we can use a set of functions that
+allow to download all the required layers from open databases. However,
+if we would like to use other input data, we can run the `copy_input`
+function which only requires the project root path, the country name,
+and the path of the file or the folder that we would like to import. To
+avoid possible further conflicts, the input is highly recommended to be
+a “raw” input (i.e. not projected). For instance, we want to use a
+country boundary shapefile from another source than the one available
+through inAccessMod. We call the function `copy_input` and indicate, as
+third parameter, the path to the shapefile folder.
+
+    country <- "Switzerland"
+    inputPath <- system.file("extdata", "swiss_boundaries_example", package="inAccessMod")
+    copy_input(mainPath, country, inputPath)
+
+We are then asked to select the data we want to copy. In this case, as
+the folder contains shapefile of the boundaries, we would select index:
+4
+
+![](copy_inputs.png)
+
+### Download the country boundaries
+
+Let’s say that we do not have any boundary shapefile. We can download it
+with the `download_boundaries` function. This layer will help to set the
+projected coordinate reference system and will be used for cropping,
+masking and clipping the other input layers. The function requires the
[…]
