### 2023-11-13 7310636
details packages remotes
