### 2023-11-13 #5879 @ptimoner
Dear Marcelo @mponce0,

Thanks for pointing out the importance of having all installation details in the README and a quickly accessible vignette in the main repository. I followed your suggestion and made the required changes. I invite you to revise the modified [README](https://github.com/unige-geohealth/inAccessMod/blob/main/README.md) and to check the new [TUTORIAL](https://github.com/unige-geohealth/inAccessMod/blob/main/TUTORIAL.md) file in the main repository.

Thanks again,

Best regards,

Pablo
