### 2023-11-13 unige-geohealth/inAccessMod#4 @ConnorDonegan
I was able to move through the tutorial, download landcover, roads, facilities, boundaries, and population raster, and open them all up in QGIS to verify that the files are what I would expect. I think the changes to the tutorial go a long way towards helping users get started. I'm going to close this one then and continue the discussion on the main review page https://github.com/openjournals/joss-reviews/issues/5879

Thanks
