### 2023-11-13 #5879 @ConnorDonegan
@ptimoner Can you comment on the issue of automated tests?

If you don't have something already in use, can you create an automated test to guard against future issues being introduced with changes to the code or dependencies? (Not necessarily using https://testthat.r-lib.org/) The tutorial now verifies that the code runs, but I would be more confident if I could see some code that will help prevent future issues.

As a review criteria we have to consider this:

- Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
