### 2023-11-15 #5879 @ptimoner
@ConnorDonegan Thank you for your valuable insights. While our tutorial already outlines the manual steps required to verify the package's functionality, to strengthen our validation process, I've integrated automated tests for both individual functions and the complete pipeline into the package. In addition to this, I've included instructions in the README on how to execute these tests.
I hope that these enhancements will address your concerns more effectively and provide comprehensive assurance about the package's functionality.

Thanks again,

Best regards,

Pablo
