### 2023-11-17 5e6d34b
author contributions and installation
