### 2023-11-17 #5879 @ptimoner
@arfon Indeed, we added an author contribution statement and a short installation guidelines as suggested by @mponce0 in the paper. Thanks.
