### 2023-11-24 #5879 @ConnorDonegan
Hi @ptimoner I'm having trouble getting the tests to run. I think its not able to find the files. My memory of it is that the tests have to be placed in a `tests/testthat/` folder.

I see that the review criteria asks for autormated _or_ manual tests, and the new vignette meets the latter criteria. I think it would be valuable to see the new automated tests run cleanly to help catch new issues as the code base changes, although its not necessary for meeting review criteria. Thanks for your responsiveness.

### The paper

As is, it follows the JOSS style---very concise, documents the need for the software and its main functionality. I think it could benefit from

 - referencing some of the alternative methods/software for completing the same tasks (i.e., some specific GIS software)
 - possibly a figure or illustration of some kind (e.g., of the required folder/file structure it produces)
 - the references for WHO reports may need URLs or additional information
 - Note on contributor guidelines and where to go for help/issues/pull requests
