+++ tests/testthat/test-09-HeRAMS_create_hf_shapefile.R
@@ -1,5 +1,3 @@
-test_file(system.file("tests/testthat/test-initiate_project.R", package = "inAccessMod"))
-test_file(system.file("tests/testthat/test-HeRAMS_filter_hf", package = "inAccessMod"))
 test_that("test HeRAMS-create_hf_shapefile", {
   expect_equal(suppressWarnings({HeRAMS_create_hf_shapefile(temp_dir, country, mostRecentBoundaries = TRUE, 
                                           rmNA = TRUE, 
