+++ tests/testthat/test-10-process_inputs.R
@@ -1,5 +1,3 @@
-test_file(system.file("tests/testthat/test-set_projection.R", package = "inAccessMod"))
-test_file(system.file("tests/testthat/test-download_population.R", package = "inAccessMod"))
 test_that("test process_inputs", {
   expect_equal(suppressWarnings({process_inputs(temp_dir, country, selectedInputs = "All",
                               mostRecent = TRUE, 
