+++ tests/testthat/test-08-HeRAMS_filter_hf.R
@@ -1,4 +1,3 @@
-test_file(system.file("tests/testthat/test-initiate_project.R", package = "inAccessMod"))
 test_that("test HeRAMS_filter_hf", {
   expect_equal(HeRAMS_filter_hf(temp_dir, country, pathTableCode = NULL, 
                                 pathTableText = NULL, 
