+++ tests/testthat/test-07-download_osm.R
@@ -1,4 +1,3 @@
-test_file(system.file("tests/testthat/test-download_boundaries.R", package = "inAccessMod"))
 test_that("test download_osm", {
   expect_equal(download_osm(x = "roads", mainPath = temp_dir, country = country, alwaysDownload = TRUE, 
                             countryName = TRUE, 
