+++ paper.bib
@@ -18,14 +18,14 @@
 @misc{herams_str:2022,
 	title = {{HeRAMS} {Strategic} {Framework}},
 	url = {https://www.who.int/publications/m/item/herams-strategic-framework},
-	author = {{World Health Organization}},
+	author = {{WHO}},
 	year = {2022},
 }
 
 @misc{herams_annual:2022,
 	title = {{HeRAMS} {Annual} {Report} {2022}},
 	url = {https://www.who.int/publications/m/item/herams-annual-report-2021},
-	author = {{World Health Organization}},
+	author = {{WHO}},
 	year = {2022},
 }
 
