### 2023-11-29 #5879 @ptimoner
Hi @ConnorDonegan,

Thanks for your comments. I changed the path of my test files, and now it should work with devtools::test().

Regarding the further comments on the paper:

1) To our knowledge, there is no alternative software for completing the same tasks the way we do here; the downloading is usually done manually, and the GIS operations are usually also performed manually using any GIS software. For this reason we decided to keep it broad and just mention that "the different GIS operations involved in pre-processing these input layers can be time-consuming and might cause issues with data consistency and quality".

2) As suggested, we added in the paper a figure that illustrates the folder/file structure the functions produce.

3) We added a reference for HeRAMS/WHO in the paper.

4) We added contributor guidelines in the README.

Thanks for taking the time to review our paper/package and for your valuable feedback.

Best regards,

Pablo
