### 2023-12-02 #5879 @mponce0
@ConnorDonegan I'm mostly OK with all the previous points I've raised as well.
However, I haven't been able to install the package yet and test its functionality since I haven't been able to successfully install it.
I tried in two different Linux systems, one arc64-based and centos, as well as MacOS and have run into dependencies issues.
@arfon @ptimoner can you please mention in which OS and corresponding versions have you tested the package?
