### 2023-12-03 #5879 @ptimoner
@mponce0 I am sorry that you couldn't install the package. Could you figure out which dependencies are not being installed correctly ? I already had dependencies issues installing other packages, and when it happens, installing the libraries causing issues one by one usually solve the issue. In the installation process, you should be able to see which dependencies are giving errors.
I am using Windows 11 but many of my collegues are using the package on MacOS (different versions) without any problem.
