### 2023-12-04 #5879 @mponce0
Hi @ptimoner Pablo, thanks for the details.

The problem I'm having is mostly with the following dependencies: `rgeos`, `rgdal`, and `gdalUtils`.
This time around even using a Windows machine...
It looks like for instance, `rgdal` has been removed from CRAN -- see https://cran.r-project.org/web/packages/rgdal/index.html
```
Package ‘rgdal’ was removed from the CRAN repository.

Formerly available versions can be obtained from the [archive](https://cran.r-project.org/src/contrib/Archive/rgdal).

Archived on 2023-10-16 at the request of the maintainer. Consider using 'sf' or 'terra' instead.

A summary of the most recent check results can be obtained from the [check results archive](https://cran-archive.r-project.org/web/checks/2023/2023-10-16_check_results_rgdal.html).

Please use the canonical form https://cran.r-project.org/package=rgdal to link to this page.
```

This is one of main concerns with the packages as it states is the "delicate" set of dependencies.
Please don't take me wrong, I really like what you are trying to achieve here but if the package can not satisfy a more robust set of dependencies then it will just work a very niche and specific set of versions of R and packages and get outdated very quickly.
Hence my original suggestion of committing it to CRAN, CRAN runs this type of checks on a continuous basis allowing developers to detect and fix this type of issues as they occur.

At this point, I'd like to ask assistance from @arfon related at how to proceed with the review?
Should we evaluate the submission as it is although it will be applicable to very narrow configuration/versions of the R ecosystem and hence it should be explicitly stated for which version of R and the packages it depends upon or ask the author to make it work for a generic setup?
