+++ R/download_dem.R
@@ -46,8 +46,8 @@ download_dem <- function (mainPath, country, alwaysDownload = FALSE, mostRecent 
     }
   }
   border <- get_boundaries(mainPath, country, "raw", mostRecent)
-  border <- as(sf::st_combine(border), "Spatial")
-  # Download SRTM tiles shapefile in a temporary folder
+
+    # Download SRTM tiles shapefile in a temporary folder
   timeFolder <- format(Sys.time(), "%Y%m%d%H%M%S")
   pathDEM <- file.path(pathDEM, timeFolder, "raw")
   check_path_length(pathDEM)
@@ -58,21 +58,21 @@ download_dem <- function (mainPath, country, alwaysDownload = FALSE, mostRecent 
   urlSRTM <- "https://github.com/sikli/srtm_country/archive/master.zip"
   utils::download.file(url = urlSRTM, destfile = paste0(tmpFolder, "/srtm.zip"))
   utils::unzip(zipfile = paste0(tmpFolder, "/srtm.zip"), overwrite = TRUE, exdir= tmpFolder)
-  shp <- raster::shapefile(file.path(tmpFolder, "srtm_country-master", "srtm", "tiles.shp"))
+  shp <- sf::st_read(file.path(tmpFolder, "srtm_country-master", "srtm", "tiles.shp"))
+  
   # Here it does not matter if border is not in lon lat (for land cover, we have a check)
-  border <- sp::spTransform(border, shp@proj4string)
-  intersects <- rgeos::gIntersects(border, shp, byid=TRUE)
-  tiles <- shp[intersects[,1],]
+  border <- sf::st_transform(border, crs = sf::st_crs(shp))
+  intersects <- suppressWarnings(sf::st_intersects(border, shp))
+  tiles <- shp[unique(unlist(intersects)), ]
   logTxt <- file.path(mainPath, country, "data", "log.txt")
   revertDownloadOptions <- FALSE
-  #Download tiles
-  if (length(tiles) > 1) {
+  # Download tiles
+  if (nrow(tiles) > 1) {
     srtmList  <- list()
-    for (i in 1:length(tiles)) {
-      cat(paste0("Downloading tile ", i, "/", length(tiles), "...\n"))
+    for (i in 1:nrow(tiles)) {
+      cat(paste0("Downloading tile ", i, "/", nrow(tiles), "...\n"))
       lon <- raster::extent(tiles[i,])[1]  + (raster::extent(tiles[i,])[2] - raster::extent(tiles[i,])[1]) / 2
       lat <- raster::extent(tiles[i,])[3]  + (raster::extent(tiles[i,])[4] - raster::extent(tiles[i,])[3]) / 2
-      # List and write (list ready if mosaic_gdal doesn't work)
       tile <- tryCatch({geodata::elevation_3s(lon = lon, lat = lat, path = tmpFolder)}, error = function (e) NULL)
       if (is.null(tile)) {
         message("Cannot open URL. Trying with 'curl' and ignoring potential SSL issues.")
@@ -86,37 +86,23 @@ download_dem <- function (mainPath, country, alwaysDownload = FALSE, mostRecent 
       options(download.file.method=NULL, download.file.extra=NULL)
     }
     cat(paste0("Creating a mosaic with the downloaded rasters...\n"))
-    # Gdal mosaic (faster)
-    files <- list.files(tmpFolder, pattern = "tif", full.names = TRUE)
-    # Try gdal (tryCatch to avoid function to stop)
-    # mosaicGDAL <- tryCatch({gdalUtils::mosaic_rasters(gdalfile = files, dst_dataset = file.path(pathDEM, "srtm.tif"), of = "GTiff")}, error = function (e) NULL, warning = function (e) NULL)
[…]
