### 2023-12-04 #5879 @ptimoner
Hi @mponce0,

Thanks for pointing this issue out. We fully agree with you that requiring dependencies that have been removed from CRAN and that are not going to be maintained anymore is not a robust solution. For this reason, we decided to modify our codes and use 'sf' and 'terra' only for all the spatial analysis, and to discard the use of 'rgeos', 'rgdal' and 'gdalUtils'.

Please let us know if you are now able to install the package.

Thanks again,

Pablo
