### 2023-12-07 #5879 @mponce0
@ptimoner Hi Pablo,
Great news! In both systems I'm testing your packages, with the latest changes in dependencies, the installation was successful.

However, all the tests cases are failing.
```R
[ FAIL 10 | WARN 1 | SKIP 0 | PASS 1 ]
══ Terminated early ════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
```

I will open an issue in your repo with the full details of the output and   ##cases...

One more thing, that it may be obvious, but still worth mentioning is that the user must have a working internet connection in order to access more if not all of the functionalities of the package.
