### 2023-12-07 unige-geohealth/inAccessMod#5 @mponce0
failing auto-tests cases (JOSS review)
The "Automated Tests Cases" are failing. See details below,

```R
> devtools::test()
ℹ Testing inAccessMod
✔ | F W  S  OK | Context
⠏ |          0 | 01-initiate_project                                                                                                                            /tmp/RtmpoV3Hm9/Syria/data
├── config.txt
├── log.txt
├── rDEM
├── rLandcover
├── rPopulation
├── vBorders
├── vFacilities
├── vNaturalPolygons
├── vRoads
└── vWaterLines

✖ | 2        1 | 01-initiate_project
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Failure (test-01-initiate_project.R:5:3): test initiate_project
file.exists(file.path(temp_dir, country, "data", "config.txt")) is not TRUE

`actual`:   FALSE
`expected`: TRUE

Failure (test-01-initiate_project.R:6:3): test initiate_project
file.exists(file.path(temp_dir, country, "data", "log.txt")) is not TRUE

`actual`:   FALSE
`expected`: TRUE
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
⠏ |          0 | 02-download_boundaries
✖ | 1        0 | 02-download_boundaries
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-02-download_boundaries.R:2:3): test download_boundaries
Error in `check_exists(pathBorder, "raw")`: /tmp/RtmpoV3Hm9/Switzerland/data/vBorders does not exist. Run the initiate_project function first or check the inputparameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-02-download_boundaries.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_boundaries(...)
5.   └─inAccessMod::check_exists(pathBorder, "raw") at inAccessMod/R/download_boundaries.R:49:3
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
⠏ |          0 | 03-set_projection
✖ | 1 1      0 | 03-set_projection
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Warning (test-03-set_projection.R:2:3): test set_projection
cannot open file '/tmp/RtmpoV3Hm9/Switzerland/data/config.txt': No such file or directory
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-03-set_projection.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::set_projection(...)
5.   └─base::file(...) at inAccessMod/R/set_projection.R:43:3

Error (test-03-set_projection.R:2:3): test set_projection
Error in `file(file.path(mainPath, country, "data", "config.txt"), open = "r")`: cannot open the connection
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-03-set_projection.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::set_projection(...)
5.   └─base::file(...) at inAccessMod/R/set_projection.R:43:3
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
⠏ |          0 | 04-download_population
✖ | 1        0 | 04-download_population
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-04-download_population.R:2:3): test download_population
Error in `check_exists(pathPop, "raw", layer = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/rPopulation does not exist. Run the initiate_project function first or check the input parameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-04-download_population.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_population(...)
5.   └─inAccessMod::check_exists(pathPop, "raw", layer = TRUE) at inAccessMod/R/download_population.R:36:3
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
⠏ |          0 | 05-download_dem
✖ | 1        0 | 05-download_dem
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-05-download_dem.R:2:3): test download_dem
Error in `check_exists(pathDEM, "raw", layer = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/rDEM does not exist. Run the initiate_project function first or check the input parameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-05-download_dem.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_dem(...)
5.   └─inAccessMod::check_exists(pathDEM, "raw", layer = TRUE) at inAccessMod/R/download_dem.R:42:3
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
⠏ |          0 | 06-download_landcover
✖ | 1        0 | 06-download_landcover
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-06-download_landcover.R:2:3): test download_landcover
Error in `check_exists(pathLandcover, "raw", layer = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/rLandcover does not exist. Run the initiate_project function firstor check the input parameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-06-download_landcover.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_landcover(...)
5.   └─inAccessMod::check_exists(pathLandcover, "raw", layer = TRUE) at inAccessMod/R/download_landcover.R:41:3
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
⠏ |          0 | 07-download_osm
✖ | 1        0 | 07-download_osm
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-07-download_osm.R:2:3): test download_osm
Error in `check_exists(pathFolder, "raw", layer = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/vRoads does not exist. Run the initiate_project function first or check the input parameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-07-download_osm.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_osm(...)
5.   └─inAccessMod::check_exists(pathFolder, "raw", layer = TRUE) at inAccessMod/R/download_osm.R:61:3
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
⠏ |          0 | 08-HeRAMS_filter_hf
HeRAMS tables: OK

✖ | 1        0 | 08-HeRAMS_filter_hf
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-08-HeRAMS_filter_hf.R:2:3): test HeRAMS_filter_hf
Error in `HeRAMS_filter_hf(temp_dir, country, pathTableCode = NULL, pathTableText = NULL,
scenario = NULL, mostRecentObs = TRUE, defaultParameters = TRUE,
region = FALSE, type = FALSE, ownership = FALSE, status = FALSE,
building = FALSE, equipment = FALSE, functionality = FALSE,
accessibility = FALSE, support = FALSE, services = FALSE,
partners = FALSE, barriers = FALSE, testMode = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/vFacilities  does not exist. Run the initiate_project function.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-08-HeRAMS_filter_hf.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::HeRAMS_filter_hf(...)
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
⠏ |          0 | 09-HeRAMS_create_hf_shapefile
✖ | 1        0 | 09-HeRAMS_create_hf_shapefile
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-09-HeRAMS_create_hf_shapefile.R:2:34): test HeRAMS-create_hf_shapefile
Error in `HeRAMS_create_hf_shapefile(temp_dir, country, mostRecentBoundaries = TRUE,
rmNA = TRUE, rmOut = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/vFacilities  does not exist. Run the initiate_project function.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-09-HeRAMS_create_hf_shapefile.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. ├─base::suppressWarnings(...)
5. │ └─base::withCallingHandlers(...)
6. └─inAccessMod::HeRAMS_create_hf_shapefile(...) at test-09-HeRAMS_create_hf_shapefile.R:2:34
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Maximum number of failures exceeded; quitting at end of file.
ℹ Increase this number with (e.g.) `testthat::set_max_fails(Inf)`

══ Results ═════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
── Failed tests ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Failure (test-01-initiate_project.R:5:3): test initiate_project
file.exists(file.path(temp_dir, country, "data", "config.txt")) is not TRUE

`actual`:   FALSE
`expected`: TRUE

Failure (test-01-initiate_project.R:6:3): test initiate_project
file.exists(file.path(temp_dir, country, "data", "log.txt")) is not TRUE

`actual`:   FALSE
`expected`: TRUE

Error (test-02-download_boundaries.R:2:3): test download_boundaries
Error in `check_exists(pathBorder, "raw")`: /tmp/RtmpoV3Hm9/Switzerland/data/vBorders does not exist. Run the initiate_project function first or check the inputparameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-02-download_boundaries.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_boundaries(...)
5.   └─inAccessMod::check_exists(pathBorder, "raw") at inAccessMod/R/download_boundaries.R:49:3

Error (test-03-set_projection.R:2:3): test set_projection
Error in `file(file.path(mainPath, country, "data", "config.txt"), open = "r")`: cannot open the connection
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-03-set_projection.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::set_projection(...)
5.   └─base::file(...) at inAccessMod/R/set_projection.R:43:3

Error (test-04-download_population.R:2:3): test download_population
Error in `check_exists(pathPop, "raw", layer = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/rPopulation does not exist. Run the initiate_project function first or check the input parameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-04-download_population.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_population(...)
5.   └─inAccessMod::check_exists(pathPop, "raw", layer = TRUE) at inAccessMod/R/download_population.R:36:3

Error (test-05-download_dem.R:2:3): test download_dem
Error in `check_exists(pathDEM, "raw", layer = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/rDEM does not exist. Run the initiate_project function first or check the input parameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-05-download_dem.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_dem(...)
5.   └─inAccessMod::check_exists(pathDEM, "raw", layer = TRUE) at inAccessMod/R/download_dem.R:42:3

Error (test-06-download_landcover.R:2:3): test download_landcover
Error in `check_exists(pathLandcover, "raw", layer = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/rLandcover does not exist. Run the initiate_project function firstor check the input parameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-06-download_landcover.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_landcover(...)
5.   └─inAccessMod::check_exists(pathLandcover, "raw", layer = TRUE) at inAccessMod/R/download_landcover.R:41:3

Error (test-07-download_osm.R:2:3): test download_osm
Error in `check_exists(pathFolder, "raw", layer = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/vRoads does not exist. Run the initiate_project function first or check the input parameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-07-download_osm.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_osm(...)
5.   └─inAccessMod::check_exists(pathFolder, "raw", layer = TRUE) at inAccessMod/R/download_osm.R:61:3

Error (test-08-HeRAMS_filter_hf.R:2:3): test HeRAMS_filter_hf
Error in `HeRAMS_filter_hf(temp_dir, country, pathTableCode = NULL, pathTableText = NULL,
scenario = NULL, mostRecentObs = TRUE, defaultParameters = TRUE,
region = FALSE, type = FALSE, ownership = FALSE, status = FALSE,
building = FALSE, equipment = FALSE, functionality = FALSE,
accessibility = FALSE, support = FALSE, services = FALSE,
partners = FALSE, barriers = FALSE, testMode = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/vFacilities  does not exist. Run the initiate_project function.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-08-HeRAMS_filter_hf.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::HeRAMS_filter_hf(...)

Error (test-09-HeRAMS_create_hf_shapefile.R:2:34): test HeRAMS-create_hf_shapefile
Error in `HeRAMS_create_hf_shapefile(temp_dir, country, mostRecentBoundaries = TRUE,
rmNA = TRUE, rmOut = TRUE)`: /tmp/RtmpoV3Hm9/Switzerland/data/vFacilities  does not exist. Run the initiate_project function.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-09-HeRAMS_create_hf_shapefile.R:2:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. ├─base::suppressWarnings(...)
5. │ └─base::withCallingHandlers(...)
6. └─inAccessMod::HeRAMS_create_hf_shapefile(...) at test-09-HeRAMS_create_hf_shapefile.R:2:34

[ FAIL 10 | WARN 1 | SKIP 0 | PASS 1 ]
══ Terminated early ════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
>
```

Details about the R environment,
```R
> sessionInfo()
R version 4.3.2 (2023-10-31)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 22.04.3 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3
LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so;  LAPACK version 3.10.0

locale:
[1] LC_CTYPE=en_CA.UTF-8       LC_NUMERIC=C
[3] LC_TIME=en_CA.UTF-8        LC_COLLATE=en_CA.UTF-8
[5] LC_MONETARY=en_CA.UTF-8    LC_MESSAGES=en_CA.UTF-8
[7] LC_PAPER=en_CA.UTF-8       LC_NAME=C
[9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_CA.UTF-8 LC_IDENTIFICATION=C

time zone: America/Toronto
tzcode source: system (glibc)

attached base packages:

other attached packages:
[1] inAccessMod_1.0 testthat_3.2.1

[1] DBI_1.1.3               bitops_1.0-7            geodata_0.5-9
[4] remotes_2.4.2.1         writexl_1.4.2           rematch2_2.1.2
[7] readxl_1.4.3            rlang_1.1.2             magrittr_2.0.3
[10] e1071_1.7-14            compiler_4.3.2          png_0.1-8
[13] callr_3.7.3             vctrs_0.6.5             stringr_1.5.1
[16] profvis_0.3.8           httpcode_0.3.0          pkgconfig_2.0.3
[19] crayon_1.5.2            fastmap_1.1.1           ellipsis_0.3.2
[22] leafem_0.2.3            utf8_1.2.4              promises_1.2.1
[25] rmarkdown_2.25          sessioninfo_1.2.2       ps_1.7.5
[28] purrr_1.0.2             waldo_0.5.2             xfun_0.41
[31] satellite_1.0.4         cachem_1.0.8            jsonlite_1.8.8
[34] later_1.3.2             mapview_2.11.2          terra_1.7-55
[37] prettyunits_1.2.0       R6_2.5.1                stringi_1.8.2
[40] hoardr_0.5.3            pkgload_1.3.3           brio_1.1.3
[43] lubridate_1.9.3         cellranger_1.1.0        Rcpp_1.0.11
[46] rgeoboundaries_1.2.9000 knitr_1.45              usethis_2.2.2
[49] base64enc_0.1-3         httpuv_1.6.13           timechange_0.2.0
[52] tidyselect_1.2.0        rstudioapi_0.15.0       codetools_0.2-18
[55] miniUI_0.1.1.1          curl_5.1.0              processx_3.8.2
[58] osmextract_0.5.0        pkgbuild_1.4.2          crsuggest_0.4
[61] lattice_0.20-45         tibble_3.2.1            shiny_1.8.0
[64] withr_2.5.2             evaluate_0.23           desc_1.4.2
[67] sf_1.0-14               units_0.8-5             proxy_0.4-27
[70] urlchecker_1.0.1        exactextractr_0.10.0    pillar_1.9.0
[73] KernSmooth_2.23-20      stats4_4.3.2            generics_0.1.3
[76] rprojroot_2.0.4         sp_2.1-2                RCurl_1.98-1.13
[79] munsell_0.5.0           scales_1.3.0            xtable_1.8-4
[82] class_7.3-20            glue_1.6.2              tools_4.3.2
[85] diffobj_0.3.5           data.table_1.14.8       fs_1.6.3
[88] grid_4.3.2              crosstalk_1.2.1         devtools_2.4.5
[91] colorspace_2.1-0        raster_3.6-26           cli_3.6.1
[94] rappdirs_0.3.3          fasterize_1.0.5         fansi_1.0.5
[97] dplyr_1.1.4             countrycode_1.5.0       digest_0.6.33
[100] classInt_0.4-10         crul_1.4.0              htmlwidgets_1.6.4
[103] memoise_2.0.1           htmltools_0.5.7         lifecycle_1.0.4
[106] leaflet_2.2.1           httr_1.4.7              mime_0.12
>
```

https://github.com/openjournals/joss-reviews/issues/5879
