### 2023-12-07 6c68661
testMode Switzerland
