### 2023-12-07 unige-geohealth/inAccessMod#5 @ptimoner
Hi @mponce0,

Thank you very much for pointing out this issue. Because of the sequential nature of the process, as the first function failed, all the other functions failed as well. The automatic selection of the country index in the "initiate_project" function in test mode was not consistent with the country used for testing (maybe due to different version of the "countrycode" dependency). I slightly modified the code to avoid this potential inconsistency.

Let me know if you still experience an error, and if not, would you please close this issue ?

Thanks again,

Best regards,

Pablo
