### 2023-12-07 #5879 @ptimoner
Thanks @mponce0 for pointing out this issue of the failed automated tests. It has been addressed (https://github.com/unige-geohealth/inAccessMod/issues/5#issue-2030966346) and they should work now. As suggested, I also mentioned in the README and the Tutorial that the user must have a working internet connection in order to access all the functions that allow downloading data.

Best regards,

Pablo
