### 2023-12-09 39c0ceb
adjusting names in standard input values to match case-sensitive values
vWaterLines -> vWaterlines
vNaturalPolygons -> vNaturalpolygons
