### 2023-12-09 unige-geohealth/inAccessMod#5 @mponce0
Thanks this fixed most of the issues, however there is one test still failing in Linux systems: test #7

```R
── Failed tests ───────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-07-download_osm.R:6:3): test download_osm
Error in `check_exists(pathFolder, "raw", layer = TRUE)`: /tmp/RtmpAlQR50/Switzerland/data/vWaterlines does not exist. Run the initiate_project function first or check the input parameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-07-download_osm.R:6:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_osm(...)
5.   └─inAccessMod::check_exists(pathFolder, "raw", layer = TRUE) at inAccessMod/R/download_osm.R:61:3

[ FAIL 1 | WARN 0 | SKIP 0 | PASS 14 ]
```

Actually if you look into the data directory you will see the problem:
```R
> system("ls  /tmp/RtmpAlQR50/Switzerland/data/")
config.txt  rDEM        rPopulation  vFacilities       vRoads       zToAccessMod
log.txt     rLandcover  vBorders     vNaturalPolygons  vWaterLines
```

The issue is that somehow the R code refers to "vWaterlines" (notice the lower case "l") vs the "vWaterLines" (capital case "L") in the directories names.
BTW this does not happen in MacOS as it is not case sensitive, i.e. in MacOS all the tests cases are passing but not in case-sensitive OSes, such as, Linux.

Looking into your `initiate_project.R` module, there are a couple of things that could fix this by switching to "vWaterlines", as well as, "vNaturalpolygons" (for similar reasons).

I just created a pull-request for you to review with the changes, which I believe will fix this issue.
