### 2023-12-09 unige-geohealth/inAccessMod#5 @ptimoner
Thanks a lot Marcelo.

As explained here https://github.com/unige-geohealth/inAccessMod/pull/6#issuecomment-1848641079, as we had the word "waterLines" in multiple places in the package (especially in code examples), we finally decided to modify the line in `download_osm` that was causing the issue without the need of replacing "waterLines" by "waterlines" everywhere.

It might work now.

Best regards.
