### 2023-12-10 unige-geohealth/inAccessMod#5 @mponce0
Sorry, but it is failing with the same error if I try the current version from the repository.

```R
── Failed tests ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-07-download_osm.R:6:3): test download_osm
Error in `check_exists(pathFolder, "raw", layer = TRUE)`: /tmp/Rtmp344lKk/Switzerland/data/vWaterlines does not exist. Run the initiate_project function first orcheck the input parameters.
Backtrace:
▆
1. ├─testthat::expect_equal(...) at test-07-download_osm.R:6:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. └─inAccessMod::download_osm(...)
5.   └─inAccessMod::check_exists(pathFolder, "raw", layer = TRUE) at inAccessMod/R/download_osm.R:61:3

[ FAIL 1 | WARN 0 | SKIP 0 | PASS 14 ]
```

and data directory appears to be unchanged in terms of sub-directories names"
```R
> system('ls  /tmp/Rtmp344lKk/Switzerland/data')
config.txt  log.txt  rDEM  rLandcover  rPopulation  vBorders  vFacilities  vNaturalPolygons  vRoads  vWaterLines  zToAccessMod
```
