### 2023-12-10 unige-geohealth/inAccessMod#5 @mponce0
Tests are passing both in Linux and MacOS!
Thank you for looking into this, will close this issue.

I've found another issue though, I will open a separate issue for it.
