### 2023-12-10 unige-geohealth/inAccessMod#7 @mponce0
error when calling label_landcover funciton as shown in tutorial
The `label_landcover` is failing when I used as instructed in the tutorial,
```R
> label_landcover(mainPath, country, mostRecent = TRUE, overwrite = TRUE, defaultLabels = TRUE)
Error in label_landcover(mainPath, country, mostRecent = TRUE, overwrite = TRUE,  :

Processed land cover raster is missing!
```

Same error shows up when trying the example presented in the help of the label_landcover function,
```R
 > country
[1] "Switzerland"
>
> download_boundaries(mainPath, country, adminLevel = 1, type = "gbOpen", alwaysDownload = TRUE)
trying URL 'https://github.com/wmgeolab/geoBoundaries/raw/c0ed7b8/releaseData/gbOpen/CHE/ADM1/geoBoundaries-CHE-ADM1-all.zip'
Content type 'application/zip' length 5546581 bytes (5.3 MB)
==================================================
downloaded 5.3 MB

Done: /.../inAccessMod/Switzerland/data/vBorders/20231210113933/raw/geoBoundaries-CHE-ADM1.shp
[1] TRUE
>
> download_landcover(mainPath, country, alwaysDownload = TRUE, mostRecent = TRUE)Loading raw boundary shapefile...
trying URL 'https://s3-eu-west-1.amazonaws.com/vito.landcover.global/v3.0.1/2019/E000N60/E000N60_PROBAV_LC100_global_v3.0.1_2019-nrt_Discrete-Classification-map_EPSG-4326.tif'
Content type 'image/tiff' length 67306666 bytes (64.2 MB)
==================================================
downloaded 64.2 MB

Done: /.../inAccessMod/Switzerland/data/rLandcover/20231210113952/raw/Switzerland_PROBAV_LC100_global_v3.0.1_2019-nrt_Discrete-Classification-map_EPSG-4326.tif
[1] TRUE
>
> label_landcover(mainPath, country, mostRecent = TRUE, overwrite = TRUE, defaultLabels = TRUE)
Error in label_landcover(mainPath, country, mostRecent = TRUE, overwrite = TRUE,  :

Processed land cover raster is missing!
>
```

One more thing, please notice that the help for the ` label_landcover` function has a typo and a possible type mismatch in the first part of its example:
```R
Examples:

# Replace workDir with the actual path to your working directory
## Not run:

mainPath <- "workDir"
intiate_project(mainPath, country)
## End(Not run)
```

typo: `intiate_project(mainPath, country)` --> missing "i" in `initiate_project", also `country` does not look to be a parameter for the function

openjournals/joss-reviews#5879
