### 2023-12-10 6792fa6
label_landcover example
