+++ R/copy_input.R
@@ -11,7 +11,7 @@
 #' # Replace workDir with the actual path to your working directory
 #' \dontrun{
 #' mainPath <- "workDir"
-#' intiate_project(mainPath, country)}
+#' initiate_project(mainPath, country)}
 #' 
 #' # Replace myCountry with the country name you are working on (workDir subfolder)
 #' # Replace myInputPath with the actual path to the file to be copied or to the directory
