+++ R/download_population.R
@@ -14,7 +14,7 @@
 #' # Replace workDir with the actual path to your working directory
 #' \dontrun{
 #' mainPath <- "workDir"
-#' intiate_project(mainPath, country)}
+#' initiate_project(mainPath, country)}
 #' 
 #' # Replace myCountry with the country name you are working on (workDir subfolder)
 #' \dontrun{
