### 2023-12-10 unige-geohealth/inAccessMod#7 @ptimoner
@mponce0 Yes, actually the label_landcover function requires a processed landcover layer. This is why you got the message "Processed land cover raster is missing!". In the tutorial it comes after the processing of the input data. To make it clearer we added in the tutorial and in the function help a note about this. We also corrected the example in the help. Indeed, prior operations that were not included in the example are needed in order to obtain a processed landcover layer.

We also corrected the typo (thanks !) and another issue that could raise when processing the landcover layer only.

Please pull the updated version of the package and let us know if you could run the label_landcover correctly.

Thanks.

https://github.com/openjournals/joss-reviews/issues/5879
