### 2023-12-10 unige-geohealth/inAccessMod#7 @mponce0
Sorry, I'm still seeing the same error after pulling the latest commit d6178d9ee296dca4c9b4c32fd45e3eaa80bad4ef.

Also in the help the `initiative_project` still shows up the `country` argument which as far as I can see is not a valid one.

For `labels_landcover` is there something I'm missing? I'm just following the steps shown in the tutorial or the help function.
Maybe a step-by-step example here could help?
