### 2023-12-10 f9bd3f8
initiate_project(mainPath)
