+++ R/HeRAMS_filter_hf.R
@@ -30,7 +30,7 @@
 #' # Replace workDir with the actual path to your working directory
 #' \dontrun{
 #' mainPath <- "workDir"
-#' initiate_project(mainPath, country)}
+#' initiate_project(mainPath)}
 #' 
 #' # Replace myCountry with the country name you are working on (workDir subfolder)
 #' # Replace myHeRAMScodeTable with the path of the HeRAMS table that contains codes; set to NULL to use example data
