+++ R/check_inputs.R
@@ -12,7 +12,7 @@
 #' # Replace workDir with the actual path to your working directory
 #' \dontrun{
 #' mainPath <- "workDir"
-#' initiate_project(mainPath, country)}
+#' initiate_project(mainPath)}
 #' 
 #' # Replace myCountry with the country name you are working on (workDir subfolder)
 #' \dontrun{
