+++ R/label_landcover.R
@@ -21,7 +21,7 @@
 #' download_boundaries(mainPath, country, adminLevel = 1, type = "gbOpen", alwaysDownload = TRUE)
 #' set_projection(mainPath, country) #required for processing the landcover
 #' download_landcover(mainPath, country, alwaysDownload = TRUE, mostRecent = TRUE)
-#' download_population(mainPath, country, alwaysDownload = TRUE, mostRecent = TRUE) #required for processing the landcover
+#' download_population(mainPath, country, alwaysDownload = TRUE) #required for processing the landcover
 #' process_inputs(mainPath, country, "rLandcover", defaultMethods = TRUE, changeRes = FALSE, popCorrection = FALSE)
 #' label_landcover(mainPath, country, mostRecent = TRUE, overwrite = TRUE, defaultLabels = TRUE)}
 #' @export
