### 2023-12-10 unige-geohealth/inAccessMod#7 @ptimoner
@mponce0 If you still have the same error, it is likely that the something went wrong with the downloading and processing of the landcover raster. I slightly modified the help, and if you follow the different steps it should work.

1. `initiate_project(mainPath)` #mainPath is your working directory; select Switzerland
2. `country <- "Switzerland"`
3. `download_boundaries(mainPath, country, adminLevel = 1, type = "gbOpen", alwaysDownload = TRUE)`
4. `set_projection(mainPath, country, mostRecent = TRUE, alwaysSet = TRUE, bestCRS = TRUE)` # you will interactively asked to provide a label for the processed boundaries (when setting the projection, the boundary shapefile is processed)
5. `download_landcover(mainPath, country, alwaysDownload = TRUE, mostRecent = TRUE)`
6. `download_population(mainPath, country, alwaysDownload = TRUE)` # Go to Global_2000_2020_Constrained/2019 and select che_ppp_2019_constrained.tif
7. `process_inputs(mainPath, country, "rLandcover", defaultMethods = TRUE, changeRes = FALSE, popCorrection = FALSE)` # you will be interactively asked to provide a label for the landcover raster.
8. `label_landcover(mainPath, country, mostRecent = TRUE, overwrite = TRUE, defaultLabels = TRUE)`

The boundaries as well as the population raster are required to process any input data. For this reason you need to download them and to set the projection before processing the landcover raster. Once you have processed the landcover raster, you can run the `label_landcover` function.

If you still see the error, please let me know which step is causing the issue.

I removed the country argument in the `initiate_project` help, thanks for informing me about this.
