### 2023-12-11 unige-geohealth/inAccessMod#7 @mponce0
Thanks @ptimoner .
Everything appears to be working now, test cases and `label_landcover` function.
