### 2023-12-11 #5879 @mponce0
Alright, all good from my side now!
I'd still encourage the authors to consider the package to CRAN, it will definitely help streamlining the process of CD/CI.
