### 2024-01-03 #5879 @ptimoner
Dear @arfon,
I am writing to inquire about the current status of our paper. I would like to kindly request an update on the review process and whether there is any further information or actions required from our side.
Thanks a lot,
Best regards,
Pablo
