### 2024-01-07 #5879 @ptimoner
@arfon We gave our paper a final read, we did not find any typo and no change has been made.
We made a new release (v1.1), and upload the archive in Zenodo making sure that the title and the authors correspond to our paper.
Here is the DOI: 10.5281/zenodo.10466894
Here the link to the record in zenodo: https://zenodo.org/records/10466894
Thanks a lot,
Best regards,
Pablo
